Server/requirements.txt
fastapi==0.103.2
uvicorn==0.22.0
openai>=1.0.0
python-dotenv==1.0.0
gunicorn==21.2.0
aiofiles==23.2.1
pydantic==2.5.0
python-multipart
